requirements.txt
# Install PyTorch with CUDA 12.1 (Closest Officially Supported Version)
--index-url https://download.pytorch.org/whl/cu121
torch>=2.1.2+cu121
torchvision>=0.16.2+cu121
torchaudio>=2.1.2+cu121

# Hugging Face Libraries
transformers>=4.38.1
datasets>=2.18.0
peft>=0.9.0
accelerate>=0.26.0

# Tokenization & Processing
sentencepiece>=0.1.99
protobuf>=3.20.3
pyyaml>=6.0.1
tqdm>=4.66.2

# PDF Processing
pypdf2>=3.0.1
